fix missing lnaddress conf for clightning.
Not completely added to config by default yet beause there is still some
issue with clightning not supporting description hashes within invoices,
these are required by some wallets to confirm the lightning address
invoice is legitimate.

So currently the addresses can work, depending on whether the wallet
makes the check or not for that field.