Fix lnd.py
SatSate was not able to create an invoice through LND, because of a typo.

Anyway, despite being able to create invoices through LND, the LND support is still broken. In fact, if you try to send 1 sat from the web interface, SatSale will create an invoice and, after very few seconds, it will mark the invoce as paid and close the web page. All without being able to pay the invoice for real!